gmx check added
Checks if gmx is set, if not program now checks if it is available and installed. If yes, the program now automatically sets the version of gromacs to what is available on arc currently (gromacs/2022.4).

IN THE FUTURE:
The function will be modified so that the user can confirm and choose the file, in the event of multiple file versions or in the event that 2022.4 is not available. hardcoding rarely ends well.